// webui/src/proto/orchestrator_account.ts
export enum AccountEmailStrategy {
  ACCOUNT_EMAIL_STRATEGY_UNSPECIFIED = "ACCOUNT_EMAIL_STRATEGY_UNSPECIFIED",
  ACCOUNT_EMAIL_STRATEGY_EXPLICIT = "ACCOUNT_EMAIL_STRATEGY_EXPLICIT",
  ACCOUNT_EMAIL_STRATEGY_POOLED_PRIMARY = "ACCOUNT_EMAIL_STRATEGY_POOLED_PRIMARY",
  ACCOUNT_EMAIL_STRATEGY_POOLED_ALIAS = "ACCOUNT_EMAIL_STRATEGY_POOLED_ALIAS",
  UNRECOGNIZED = "UNRECOGNIZED",
}
// ...
export interface CreateGPTAccountRequest {
  email: string;
  password: string;
  account_id: string;
  email_strategy: AccountEmailStrategy;
  country_code: string;
  region: string;
}

export interface CreateGPTAccountResponse {
  account: Account | undefined;
  error_message: string;
}